backend/requirements.txt
Flask==2.3.2
pandas==2.0.3
python-dotenv==1.0.0
openai==0.27.8
flask-cors==3.0.10